:recycle:refactor: auth 도메인 분리 및 패키지 구조 개선
